Refactor postgres database and add warning logs

+++ packages/core/src/database/postgres/service.ts
@@ -286,40 +286,53 @@ export class PostgresDatabaseService implements BaseDatabaseService {
             return { status: "success", checkpoint: zeroCheckpoint } as const;
           }
 
-          if (namespaceLockRow === undefined) {
-            // If the lock row is held and has not expired, we cannot proceed.
-            const expiresAt =
-              buildIdLockRow!.heartbeat_at +
-              this.common.options.databaseHeartbeatTimeout;
+          // When a row is found for both types of lock rows, build ID takes
+          // precedence because its more likely to skip re-indexing.
+          // Note: build ID and namespace can be referring to the same row.
+          const previousLockRow =
+            buildIdLockRow !== undefined ? buildIdLockRow : namespaceLockRow!;
 
-            if (buildIdLockRow!.is_locked === 1 && Date.now() <= expiresAt) {
-              const expiresInMs = expiresAt - Date.now();
-              return { status: "locked", expiresInMs } as const;
-            }
+          // If the lock row is held and has not expired, we cannot proceed.
+          const expiresAt =
+            previousLockRow.heartbeat_at +
+            this.common.options.databaseHeartbeatTimeout;
 
-            // If the lock row has a non-zero finalized checkpoint,
-            // we can revert unfinalized rows and continue where it left off.
-            // Note that the userNamespace is not the same, so we must switch
-            // the current app to use the namespace of the app that last had the lock.
+          if (previousLockRow.is_locked === 1 && Date.now() <= expiresAt) {
+            const expiresInMs = expiresAt - Date.now();
+            return { status: "locked", expiresInMs } as const;
+          }
+
+          // If the lock row has the same build ID as the current app AND
+          // has a non-zero finalized checkpoint, we can revert unfinalized
+          // rows and continue where it left off.
+          if (
+            this.common.options.command === "start" &&
+            previousLockRow.build_id === this.buildId &&
+            previousLockRow.finalized_checkpoint !==
+              encodeCheckpoint(zeroCheckpoint)
+          ) {
             if (
-              this.common.options.command === "start" &&
-              buildIdLockRow!.finalized_checkpoint !==
-                encodeCheckpoint(zeroCheckpoint)
+              buildIdLockRow !== undefined &&
+              buildIdLockRow.namespace !== this.userNamespace
             ) {
               this.common.logger.info({
                 service: "database",
                 msg: `Detected cache hit for build '${
                   this.buildId
                 }' in schema '${
-                  buildIdLockRow!.namespace
+                  buildIdLockRow.namespace
                 }' last active ${formatEta(
-                  Date.now() - buildIdLockRow!.heartbeat_at,
+                  Date.now() - previousLockRow.heartbeat_at,
                 )} ago`,
+              });
+              this.common.logger.warn({
+                service: "database",
+                msg: `Switching schema for indexed tables from '${this.userNamespace}' to '${buildIdLockRow.namespace}'`,
               });
 
               // Update user namespace to the namespace of the previous lock holder.
-              this.userNamespace = buildIdLockRow!.namespace;
-              namespaceInfo.userNamespace = buildIdLockRow!.namespace;
+              this.userNamespace = buildIdLockRow.namespace;
+              namespaceInfo.userNamespace = buildIdLockRow.namespace;
               namespaceInfo.internalTableIds = Object.keys(
                 getTables(schema),
               ).reduce((acc, tableName) => {
@@ -330,155 +343,16 @@ export class PostgresDatabaseService implements BaseDatabaseService {
                 ]);
                 return acc;
               }, {} as { [tableName: string]: string });
-
-              // Remove any indexes, will be recreated once the app
-              // becomes healthy.
-              for (const [tableName, table] of Object.entries(
-                getTables(schema),
-              )) {
-                if (table.constraints === undefined) continue;
-
-                for (const name of Object.keys(table.constraints)) {
-                  await tx.schema
-                    .withSchema(this.userNamespace)
-                    .dropIndex(`${tableName}_${name}`)
-                    .ifExists()
-                    .execute();
-
-                  this.common.logger.info({
-                    service: "database",
-                    msg: `Dropped index '${tableName}_${name}' in schema '${this.userNamespace}'`,
-                  });
-                }
-              }
-
-              await tx
-                .withSchema(this.internalNamespace)
-                .updateTable("namespace_lock")
-                .set({ is_locked: 1, heartbeat_at: Date.now() })
-                .where("build_id", "=", buildId)
-                .execute();
-              this.common.logger.debug({
-                service: "database",
-                msg: `Acquired lock on schema '${this.userNamespace}'`,
-              });
-
-              const finalizedCheckpoint = decodeCheckpoint(
-                buildIdLockRow!.finalized_checkpoint,
-              );
-
+            } else {
               this.common.logger.info({
                 service: "database",
-                msg: `Reverting operations prior to finalized checkpoint (timestamp=${finalizedCheckpoint.blockTimestamp} chainId=${finalizedCheckpoint.chainId} block=${finalizedCheckpoint.blockNumber})`,
+                msg: `Detected cache hit for build '${
+                  this.buildId
+                }' in schema '${this.userNamespace}' last active ${formatEta(
+                  Date.now() - previousLockRow.heartbeat_at,
+                )} ago`,
               });
-
-              // Revert unfinalized data from the existing tables.
-              const tx_ = tx as KyselyTransaction<any>;
-              for (const [tableName, tableId] of Object.entries(
-                namespaceInfo.internalTableIds,
-              )) {
-                const rows = await tx_
-                  .withSchema(namespaceInfo.internalNamespace)
-                  .deleteFrom(tableId)
-                  .returningAll()
-                  .where(
-                    "checkpoint",
-                    ">",
-                    buildIdLockRow!.finalized_checkpoint,
-                  )
-                  .execute();
-
-                const reversed = rows.sort(
-                  (a, b) => b.operation_id - a.operation_id,
-                );
-
-                for (const log of reversed) {
-                  if (log.operation === 0) {
-                    // Create
-                    await tx_
-                      .withSchema(namespaceInfo.userNamespace)
-                      .deleteFrom(tableName)
-                      .where("id", "=", log.id)
-                      .execute();
-                  } else if (log.operation === 1) {
-                    // Update
-                    log.operation_id = undefined;
-                    log.checkpoint = undefined;
-                    log.operation = undefined;
-                    await tx_
-                      .withSchema(namespaceInfo.userNamespace)
-                      .updateTable(tableName)
-                      .set(log)
-                      .where("id", "=", log.id)
-                      .execute();
-                  } else {
-                    // Delete
-                    log.operation_id = undefined;
-                    log.checkpoint = undefined;
-                    log.operation = undefined;
-                    await tx_
-                      .withSchema(namespaceInfo.userNamespace)
-                      .insertInto(tableName)
-                      .values(log)
-                      .execute();
-                  }
-                }
-
-                this.common.logger.info({
-                  service: "database",
-                  msg: `Reverted ${rows.length} unfinalized operations from existing '${tableName}' table`,
-                });
-              }
-
-              return {
-                status: "success",
-                checkpoint: finalizedCheckpoint,
-              } as const;
             }
-
-            // Otherwise, create new tables.
-            await tx
-              .withSchema(this.internalNamespace)
-              .insertInto("namespace_lock")
-              .values(newLockRow)
-              .execute();
-            this.common.logger.debug({
-              service: "database",
-              msg: `Acquired lock on new schema '${this.userNamespace}'`,
-            });
-
-            await createTables();
-
-            return { status: "success", checkpoint: zeroCheckpoint } as const;
-          }
-
-          // If the lock row is held and has not expired, we cannot proceed.
-          const expiresAt =
-            namespaceLockRow.heartbeat_at +
-            this.common.options.databaseHeartbeatTimeout;
-
-          if (namespaceLockRow.is_locked === 1 && Date.now() <= expiresAt) {
-            const expiresInMs = expiresAt - Date.now();
-            return { status: "locked", expiresInMs } as const;
-          }
-
-          // If the lock row has the same build ID as the current app AND
-          // has a non-zero finalized checkpoint, we can revert unfinalized
-          // rows and continue where it left off.
-          if (
-            this.common.options.command === "start" &&
-            namespaceLockRow.build_id === this.buildId &&
-            namespaceLockRow.finalized_checkpoint !==
-              encodeCheckpoint(zeroCheckpoint)
-          ) {
-            this.common.logger.info({
-              service: "database",
-              msg: `Detected cache hit for build '${this.buildId}' in schema '${
-                this.userNamespace
-              }' last active ${formatEta(
-                Date.now() - namespaceLockRow.heartbeat_at,
-              )} ago`,
-            });
 
             // Remove any indexes, will be recreated once the app
             // becomes healthy.
@@ -505,7 +379,7 @@ export class PostgresDatabaseService implements BaseDatabaseService {
               .withSchema(this.internalNamespace)
               .updateTable("namespace_lock")
               .set({ is_locked: 1, heartbeat_at: Date.now() })
-              .where("namespace", "=", this.userNamespace)
+              .where("build_id", "=", buildId)
               .execute();
             this.common.logger.debug({
               service: "database",
@@ -513,12 +387,12 @@ export class PostgresDatabaseService implements BaseDatabaseService {
             });
 
             const finalizedCheckpoint = decodeCheckpoint(
-              namespaceLockRow.finalized_checkpoint,
+              buildIdLockRow!.finalized_checkpoint,
             );
 
             this.common.logger.info({
               service: "database",
-              msg: `Reverting operations prior to finalized checkpoint (timestamp=${finalizedCheckpoint.blockTimestamp} chainId=${finalizedCheckpoint.chainId} block=${finalizedCheckpoint.blockNumber})`,
+              msg: `Reverting operations after the finalized checkpoint (timestamp=${finalizedCheckpoint.blockTimestamp} chainId=${finalizedCheckpoint.chainId} block=${finalizedCheckpoint.blockNumber})`,
             });
 
             // Revert unfinalized data from the existing tables.
@@ -530,7 +404,7 @@ export class PostgresDatabaseService implements BaseDatabaseService {
                 .withSchema(namespaceInfo.internalNamespace)
                 .deleteFrom(tableId)
                 .returningAll()
-                .where("checkpoint", ">", namespaceLockRow.finalized_checkpoint)
+                .where("checkpoint", ">", buildIdLockRow!.finalized_checkpoint)
                 .execute();
 
               const reversed = rows.sort(
@@ -583,10 +457,10 @@ export class PostgresDatabaseService implements BaseDatabaseService {
 
           // Otherwise, the lock row has a different build ID or a zero finalized checkpoint,
           // so we need to drop the previous app's tables and create new ones.
-          const previousBuildId = namespaceLockRow.build_id;
+          const previousBuildId = previousLockRow.build_id;
           // Note: `previousSchema` should only be used to get table names or enum names because
           // the types of `Table` and `Enum` have changed between versions.
-          const previousSchema = namespaceLockRow.schema as unknown as {
+          const previousSchema = previousLockRow.schema as unknown as {
             tables: { [tableName: string]: Table };
             enums: { [enumName: string]: Enum };
           };
@@ -603,30 +477,33 @@ export class PostgresDatabaseService implements BaseDatabaseService {
             msg: `Acquired lock on schema '${this.userNamespace}' previously used by build '${previousBuildId}'`,
           });
 
-          for (const tableName of Object.keys(previousSchema.tables)) {
-            const tableId = hash([
-              this.userNamespace,
-              previousBuildId,
-              tableName,
-            ]);
+          // Only drop tables if the previous app is using the same namespace.
+          if (namespaceLockRow !== undefined) {
+            for (const tableName of Object.keys(previousSchema.tables)) {
+              const tableId = hash([
+                this.userNamespace,
+                previousBuildId,
+                tableName,
+              ]);
 
-            await tx.schema
-              .withSchema(this.internalNamespace)
-              .dropTable(tableId)
-              .ifExists()
-              .execute();
+              await tx.schema
+                .withSchema(this.internalNamespace)
+                .dropTable(tableId)
+                .ifExists()
+                .execute();
 
-            await tx.schema
-              .withSchema(this.userNamespace)
-              .dropTable(tableName)
-              .cascade() // Need cascade here to drop dependent published views.
-              .ifExists()
-              .execute();
+              await tx.schema
+                .withSchema(this.userNamespace)
+                .dropTable(tableName)
+                .cascade() // Need cascade here to drop dependent published views.
+                .ifExists()
+                .execute();
 
-            this.common.logger.debug({
-              service: "database",
-              msg: `Dropped '${tableName}' table left by previous build`,
-            });
+              this.common.logger.debug({
+                service: "database",
+                msg: `Dropped '${tableName}' table left by previous build`,
+              });
+            }
           }
 
           await createTables();
